pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = ""
authors = ["Arnary <anastasia.dushka1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
fastapi = {extras = ["standart"], version = "^0.115.6"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.7.0"
slowapi = "^0.1.9"
uvicorn = "^0.34.0"
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.2"
cloudinary = "^1.41.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = "^5.2.1"
jsonpickle = "^4.0.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"
aiosqlite = "^0.20.0"

[tool.poetry.group.docs.dependencies]
sphinx = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

package-mode = false

[tool.pytest.ini_options]
pythonpath = ["."]
filterwarnings = "ignore::DeprecationWarning"
asyncio_default_fixture_loop_scope = "function"
